SendTweet Bug Fix & GetMinRetweets Update
SendTweet had a bug when the tweet buffers were empty and the max time
elapsed, this has been fixed. GetMinRetweets needed to be updated to
ignore outliers (very large or very small retweet counts), this was done
by first getting a standard deviation and then ignoring anything outside
of 2 standard deviations. The idea is that popular tweets will tend to
follow a bell curve and outliers can be ignored.