Repository Updates
The overall complexity of the Repository currently being used throughout
the Postworthy project has caused me to begin to refactor the repository
into a simpler design. Instead of a 3 level caching design i am working
on just a 2 level design where all data is stored in the long term
storage and a secondary cache is used to make the queries run faster.