DistributedSharedCache Bug Fix
There was a bug in the distributed cache that would seem to show up when
the site was in heavy use. I believe what was happening is it was
possible for chunked data to go get partially cleared out by the
memcache server and this would result in an error. The solution was to
just try catch the block that did deserialization of the cached data.