Tweet Count Bypass
Twitter has discontinued the api by which we can know the number of
times a url has been retweeted so we are going to have to bypass this
for now. Unfortunately this was a big part of how the postworthy
software actually ranked articles so for now we are just going to say
that all tweets are equal at 100 retweets.