RealtimeProcesingStep Upgrade
The RealtimeProcessingStep is now going to be the main way that I
personally process the stream for my sites. Due to the fact that when
you run a large number of sites the batch processing can be very slow
and problematic due to the time required to constantly pull each tweet
and process it this processing step will keep track of the top tweets in
real time so that the sites can be more responsive.

The RealtimeProcesingStep now can do everything that was once reserved
for the WebContent task.